- Modified XTEA to use uint32_t instead of unsigned long

+++ include/polarssl/xtea.h
@@ -20,6 +20,8 @@
 #ifndef POLARSSL_XTEA_H
 #define POLARSSL_XTEA_H
 
+#include <inttypes.h>
+
 #define XTEA_ENCRYPT     1
 #define XTEA_DECRYPT     0
 
@@ -29,7 +31,7 @@
  */
 typedef struct
 {
-    unsigned long k[4];       /*!< key */
+    uint32_t k[4];       /*!< key */
 }
 xtea_context;
